backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
sqlalchemy==1.4.23
python-multipart==0.0.5
python-jose==3.3.0
passlib==1.7.4
python-dotenv==0.19.0
alembic==1.7.3
pytest==6.2.5
aiofiles==0.8.0
python-multipart==0.0.5
Pillow==9.0.0
pydantic[email]>=1.8.0
email-validator>=1.1.3